[sandbox] Allow maliit by default. Fixes JB#51408
As maliit creates an abstract socket it's not visible to new namespace.
Let's not create a new namespace when there is no internet access, at
least until there is a better solution.

In any case we do not create a network namespace if networking is
allowed.

Fix that typo in profile.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>